backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-environ==0.11.2
psycopg2-binary==2.9.7
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.5.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
drf-yasg==1.21.7 